Extract Kind methods
