+++ source/app/assets/javascripts/controllers/category/controller.js
@@ -1,37 +1,8 @@
 (function(_, angular) {
   var app = angular.module("category/controller", [
-    "cyberhawk/builder"
+    "cyberhawk/builder",
+    "kind/methods"
   ]);
-
-  var KindMethods = {
-    requestKinds: function() {
-      var promise = this._getKindsRequester().request();
-      promise.then(this._setKinds);
-
-      this.constructor.trigger(this, this.route, "request");
-    },
-
-    _setKinds: function(response) {
-      this.kinds = response.data;
-      this.loaded = true;
-      this.constructor.trigger(this, this.route, "loaded");
-    },
-
-    _getKindsRequester: function() {
-      if ( !this.kindsRequester ) {
-        this._buildKindsRequester();
-      }
-
-      return this.kindsRequester;
-    },
-    _buildKindsRequester: function() {
-      this.kindsRequester = this.requesterBuilder.build({
-        search: this.location.$$search,
-        path: "/kinds"
-      });
-      this.kindsRequester.bind(this);
-    }
-  };
 
   var SubscriptionsMethods = {
     subscribe: function(categorySlug) {
@@ -49,18 +20,17 @@
   var options = {
     callback: function() {
       _.extend(this, SubscriptionsMethods);
-      _.extend(this, KindMethods);
-      _.bindAll(this, "requestKinds", "_setKinds", "_getKindsRequester", "_buildKindsRequester");
 
       _.bindAll(this, "subscribe", "_getSubscriptionRequester");
-
-      this.requestKinds();
     }
   };
 
   app.controller("Category.Controller", [
-    "cyberhawk_builder", function(builder) {
+    "cyberhawk_builder",
+    "kinds_methods",
+    function(builder, kindsMethods) {
       builder.buildAndRequest(this, options);
+      kindsMethods.build(this);
     }
   ]);
 }(window._, window.angular));
