+++ source/app/assets/javascripts/modules/kind.js
@@ -1,0 +1,7 @@
+(function(angular) {
+  angular.module("kind", [
+    "kind/methods"
+  ])
+}(window.angular));
+
+
