+++ source/app/assets/javascripts/modules/oak.js
@@ -7,6 +7,7 @@
     'login',
     'menu',
     'category',
+    'kind',
     'item'
   ]);
